:bug: Switched base field for FiltersField to DictField
Basing it on serializers.RelatedField caused the API spec generator
to determine the incorrect type (said filters: string). Basing it on
DictField is properly picked up in the API documentation, and the
functionality remains the same (tests still pass).

Additionally, marked the create/update methods as atomic transactions,
since we want the entire action to succeed or fail, and we cannot
accept an in-between error state in case something goes wrong.